[Fix] 급등 코인 리스트에서 코인 클릭시 종목 활성화까지 적용
