+++ components/vision/rised-coin/WeeklyRised.jsx
@@ -1,7 +1,7 @@
 import { useDispatch } from 'react-redux';
 import { useRouter } from 'next/router';
 import { memo, useMemo } from 'react';
-import { setKeyword } from '@/utils/redux/chartSlice';
+import { setCode, setKeyword } from '@/utils/redux/chartSlice';
 import { LinearProgress } from '@mui/material';
 import { useWeeklyTopQuery } from '@/hooks/useWeeklyTopQuery';
 import { VisionSubTitle } from '@/defaultTheme';
@@ -15,9 +15,11 @@ function WeeklyRised({ marketCodes }) {
 
   const codeMap = useMemo(() => {
     const map = {};
-    marketCodes.forEach(item => {
-      map[item.market] = item.korean_name;
-    });
+    marketCodes
+      .filter(code => code.market.includes('KRW'))
+      .forEach(item => {
+        map[item.market] = item.korean_name;
+      });
     return map;
   }, [marketCodes]);
 
@@ -44,9 +46,17 @@ function WeeklyRised({ marketCodes }) {
       .slice(0, 10); // Top 10 선정
   }, [tickers, weeklyCandles, codeMap]);
 
-  const handleKeywordSearch = coinName => {
+  const handleClickCoin = coinName => {
     dispatch(setKeyword(coinName));
-    router.push('/trade/chart');
+    const marketCode = Object.entries(codeMap).find(
+      ([code, name]) => name === coinName,
+    )?.[0];
+    if (marketCode) {
+      dispatch(setCode(marketCode));
+      setTimeout(() => {
+        router.push('/trade/chart');
+      }, 100);
+    }
   };
 
   if (isLoading) return <LinearProgress color="primary" />;
@@ -65,7 +75,7 @@ function WeeklyRised({ marketCodes }) {
                 <span className="w-8 text-left font-ng">{i + 1}</span>
                 <span
                   className="flex-1 font-ng font-bold text-left truncate cursor-pointer"
-                  onClick={() => handleKeywordSearch(coin.name)}
+                  onClick={() => handleClickCoin(coin.name)}
                 >
                   {coin.name}
                 </span>
